requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
diff-match-patch==20121119
Django==2.0.7
django-import-export==1.0.1
django-rest-swagger==2.2.0
djangorestframework==3.8.2
et-xmlfile==1.0.1
idna==2.7
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.12
odfpy==1.3.6
openapi-codec==1.3.2
openpyxl==2.5.4
pbr==4.1.0
pytz==2018.5
PyYAML==3.13
requests==2.19.1
simplejson==3.16.0
six==1.11.0
stevedore==1.28.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
xlrd==1.1.0
xlwt==1.3.0
